refactor!: mm module
1. 重写 buddy 实现，合并 kbud 与 ubud
2. 废除定长数组实现的内核态内存块分配算法
3. 内核按 multiboot 标准获取内存布局取代 fmi 结构
4. 废弃硬编码物理内存大小，内存页面独享数组动态分配
5. 废弃大小内核的区分
6. 调整内存页面结构，所有内存分配信息可直接通过页面对象获取
7. 调整内存布局，着重调整页表与 mmap 页面
8. mm 模块划入独立的目录
9. 修复超过 64MiB 时系统启动崩溃的问题

+++ fs/devfs/devfs.c
@@ -88,7 +88,7 @@ int unregister_device(int dev) {
         spinlock_acquire(&devices_lock);
         list_remove(&dev_struct->dev_list);
         spinlock_release(&devices_lock);
-        kern_kfree((u32)dev_struct);
+        kern_kfree(dev_struct);
     }
     return 0;
 }
